+++ tests/script_snapshot_tests.rs
@@ -1536,6 +1536,7 @@ script_test!(
 );
 script_test!(script_input_loop, "input_loop.wls");
 script_test!(script_read_from_command, "read_from_command.wls");
+script_test!(script_write_to_command, "write_to_command.wls");
 script_test!(script_a_b, "a_b.wls");
 script_test!(script_dynamic_variable_names, "dynamic_variable_names.wls");
 script_test!(script_mad_libs, "mad_libs.wls");